procurement_simulator/requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
openpyxl>=3.1
streamlit>=1.39
plotly>=5.18
